+++ tibiawikisql/__init__.py
@@ -19,6 +19,3 @@ __copyright__ = "Copyright 2024 Allan Galarza"
 
 __license__ = "Apache 2.0"
 __version__ = "7.0.0-alpha"
-
-from tibiawikisql import models
-from tibiawikisql.api import Article, Image, WikiClient, WikiEntry